finish 1.0

+++ src/main/java/com/kurtlike/webserv/MathLab3/GoodOldIO.java
@@ -11,6 +11,6 @@ public interface GoodOldIO {
     public void setAccuracy(double accuracy);
     public void setStartingNumberOfIterations(int n);
     public double getAnswer();
-    public double getNumberOfDivisions();
+    public int getNumberOfDivisions();
 
 }
